+++ libs/application/templates/financial-statements-inao/src/fields/IndividualElectionOperatingIncome/expenses.tsx
