+++ libs/application/templates/financial-statements-inao/src/fields/IndividualElectionOperatingIncome/income.tsx
