+++ libs/application/templates/financial-statements-inao/src/fields/IndividualElectionOperatingIncome/styles.css.ts
