--- libs/application/templates/financial-statements-inao/src/fields/PersonalElectionEquties/index.ts
@@ -1,1 +1,0 @@
-export { PersonalElectionEquities } from './PersonalElectionEquities'
