--- libs/application/templates/financial-statements-inao/src/fields/PersonalElectionOperatingIncome/index.ts
@@ -1,1 +1,0 @@
-export { PersonalElectionOperatingIncome } from './PersonalElectionOperatingIncome'
