+++ libs/application/templates/financial-statements-inao/src/forms/application/shared/keyNumbers/keyNumbersSection.ts
@@ -26,7 +26,7 @@ export const keyNumbersSection = buildSection({
           id: 'income',
           title: m.keyNumbersIncomeAndExpenses,
           description: m.fillOutAppopriate,
-          component: 'PersonalElectionOperatingIncome',
+          component: 'IndividualElectionOperatingIncome',
           childInputIds: Object.values(OPERATIONIDS),
         }),
       ],
@@ -39,7 +39,7 @@ export const keyNumbersSection = buildSection({
           id: 'equitiesAndLiabilities',
           title: m.keyNumbersDebt,
           description: m.fillOutAppopriate,
-          component: 'PersonalElectionEquities',
+          component: 'IndividualElectionEquities',
           childInputIds: Object.values(EQUITIESANDLIABILITIESIDS),
         }),
       ],
